+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fixed some statics in plugins not being patched
 - Fixed only one model of two in armor and armor addon records being patched
 - Fixed localization breaking for overridden records in ParallaxGen.esp
+- Fixed duplicate meshes not being created past PG1
+- Fixed vanilla texture sets not being included in duplicate mesh consideration
 
 ## [0.8.3] - 2025-01-22